requirements.txt
pony
redis
foursquare
python-dotenv
wikipedia
rq